requirements.txt
tweepy
prettyconf
typer
loguru
pyyaml
emoji
twikenizer
rich
